# apps/material_agent_service/openapi.yaml
openapi: 3.1.0
info:
  title: Material Agent Service
  description: REST API for VLM-based material assignment to 3D USD objects.
  version: "0.3.0"
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /assets/{session_id}/input-render:
    get:
      summary: Get input USD render (before material assignment)
      operationId: getInputRender
      tags: [assets]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: PNG preview image
          content:
            image/png:
              schema: { type: string, format: binary }
        "404": { description: Not available }
        "424": { description: Render failed }
        "503": { description: Render still in progress }
  /assets/{session_id}/generated-ref:
    get:
      summary: Get generated reference image
      operationId: getGeneratedReferenceImage
      tags: [assets]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: PNG generated reference image
          content:
            image/png:
              schema: { type: string, format: binary }
        "404": { description: Not available }
  /assets/{session_id}/generated-ref/{reference_id}:
    get:
      summary: Get generated reference image by ID
      operationId: getGeneratedReferenceImageById
      tags: [assets]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
        - name: reference_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: PNG generated reference image
          content:
            image/png:
              schema: { type: string, format: binary }
        "404": { description: Not available }
  /assets/{session_id}/previews:
    get:
      summary: List all preview images
      operationId: listPreviews
      tags: [assets]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Preview image list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PreviewList" }
        "404": { description: Session not found }
  /assets/{session_id}/preview/{image_name}:
    get:
      summary: Get a specific preview image
      operationId: getPreview
      tags: [assets]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
        - name: image_name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: PNG preview image
          content:
            image/png:
              schema: { type: string, format: binary }
        "404": { description: Not available }
  /assets/{session_id}/references:
    get:
      summary: List reference images
      operationId: listReferences
      tags: [assets]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Reference image list
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id: { type: string }
                  references:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        url: { type: string }
                  total: { type: integer }
        "404": { description: Session not found }
  /assets/{session_id}/reference/{image_name}:
    get:
      summary: Get a reference image
      operationId: getReference
      tags: [assets]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
        - name: image_name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Reference image
          content:
            image/png:
              schema: { type: string, format: binary }
        "404": { description: Not available }
  /assets/{session_id}/reference-pdfs:
    get:
      summary: List reference PDFs
      operationId: listReferencePdfs
      tags: [assets]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Reference PDF list
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id: { type: string }
                  pdfs:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        url: { type: string }
                  total: { type: integer }
        "404": { description: Session not found }
  /assets/{session_id}/reference-pdf/{pdf_name}:
    get:
      summary: Get a reference PDF
      operationId: getReferencePdf
      tags: [assets]
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
        - name: pdf_name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "404": { description: Not available }
...
components:
  schemas:
...
    PreviewList:
      type: object
      properties:
        session_id: { type: string }
        previews:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              url: { type: string }
              prim_path: { type: string, nullable: true }
              created_at: { type: string, format: date-time }
        total: { type: integer }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Optional Bearer token (or set MATERIAL_AGENT_TOKEN env var)
security:
  - {}
  - bearerAuth: []